feature(android): add support for badge count
Android does not natively support badge counters, so we utilize shortcut badger to be able to show badge counters.
To trigger these updates one needs to send a "badge" parameter in the "data" payload of a push notification. If the app is dead (as opposed to just in the background, one cannot send the "notification" payload in the push notification as the OS then takes over control.